+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.8.0 / 2016 Jan 15
+
+> Add extention class to extend the session object for simplified manipulation of the server's registered users
+
+* **Add** - `UserUtils.cs`
+* **Add** - 'GuaranteedRate.Examples.UserUtils' for `UserUtils.cs`
+
 ## v1.7.1 / 2015 Nov 19
 
 > Add Function to get list of reportable virtual fields